# server/src/swagger_output.json
{
...
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/startInsertConstraints/": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organization": {
                  "example": "any"
                },
                "month": {
                  "example": "any"
                },
                "year": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/publishBoard/": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organization": {
                  "example": "any"
                },
                "month": {
                  "example": "any"
                },
                "year": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/openToConstraints": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/employessAssigned/": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organization": {
                  "example": "any"
                },
                "month": {
                  "example": "any"
                },
                "year": {
                  "example": "any"
                },
                "changeTo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/employessAssigned": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "organization",
            "in": "query",
            "type": "string"
          },
          {
            "name": "month",
            "in": "query",
            "type": "string"
          },
          {
            "name": "year",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/schedule/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/schedule/startInsertConstraints/": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organization": {
                  "example": "any"
                },
                "month": {
                  "example": "any"
                },
                "year": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/schedule/publishBoard/": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organization": {
                  "example": "any"
                },
                "month": {
                  "example": "any"
                },
                "year": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/schedule/openToConstraints": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/schedule/employessAssigned/": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organization": {
                  "example": "any"
                },
                "month": {
                  "example": "any"
                },
                "year": {
                  "example": "any"
                },
                "changeTo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/schedule/employessAssigned": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "organization",
            "in": "query",
            "type": "string"
          },
          {
            "name": "month",
            "in": "query",
            "type": "string"
          },
          {
            "name": "year",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}